Revert "LPD-51094 Delete null checking and add line break because order matters"
This reverts commit 152c82bfd34f40875d5315705cf8723a1a534e3a.

+++ modules/apps/sharing/sharing-test/src/testIntegration/java/com/liferay/sharing/service/test/SharingEntryLocalServiceTest.java
@@ -794,10 +794,11 @@ public class SharingEntryLocalServiceTest {
 					SharingEntryAction.UPDATE));
 		}
 		finally {
-			_userGroupLocalService.deleteUserUserGroup(
-				_toUser.getUserId(), userGroup);
-
-			_userGroupLocalService.deleteUserGroup(userGroup);
+			if (userGroup != null) {
+				_userGroupLocalService.deleteUserUserGroup(
+					_toUser.getUserId(), userGroup);
+				_userGroupLocalService.deleteUserGroup(userGroup);
+			}
 		}
 	}
 
@@ -887,10 +888,11 @@ public class SharingEntryLocalServiceTest {
 					SharingEntryAction.VIEW));
 		}
 		finally {
-			_userGroupLocalService.deleteUserUserGroup(
-				_toUser.getUserId(), userGroup);
-
-			_userGroupLocalService.deleteUserGroup(userGroup);
+			if (userGroup != null) {
+				_userGroupLocalService.deleteUserUserGroup(
+					_toUser.getUserId(), userGroup);
+				_userGroupLocalService.deleteUserGroup(userGroup);
+			}
 		}
 	}
